+++ laktory/dlt/__init__.py
@@ -1,5 +1,6 @@
+from typing import TYPE_CHECKING
+
 from laktory._logger import get_logger
-from laktory.spark import SparkDataFrame
 
 try:
     from dlt import *
@@ -9,6 +10,8 @@ try:
 except Exception:
     pass
 
+if TYPE_CHECKING:
+    from pyspark.sql import DataFrame
 
 logger = get_logger(__name__)
 
@@ -63,7 +66,7 @@ def is_debug() -> bool:
     return False
 
 
-def get_df(df_wrapper) -> SparkDataFrame:
+def get_df(df_wrapper) -> "DataFrame":
     """
     When executed in debug mode (ref `dlt.is_debug`), executes and returns
     the output dataframe of a function decorated with `@dlt.table` or
@@ -112,7 +115,7 @@ def get_df(df_wrapper) -> SparkDataFrame:
 # --------------------------------------------------------------------------- #
 
 
-def read(*args, **kwargs) -> SparkDataFrame:
+def read(*args, **kwargs) -> "DataFrame":
     """
     When `is_debug()` is `True` read table from storage, else read table from
     pipeline with native Databricks `dlt.read`
@@ -242,7 +245,7 @@ def apply_changes(*args, node=None, **kwargs):
     if is_debug():
         if node is None:
             return
-        df = node.source.read(spark=spark)
+        df = node.source.read().to_native()
         # TODO: Apply changes
         logger.warning(
             "Laktory does not currently support applying CDC changes. Returned dataframe is CDC source."