+++ laktory/models/pipeline/pipelinenode.py
@@ -24,7 +24,7 @@ from laktory.models.datasources import DataSourcesUnion
 from laktory.models.datasources import PipelineNodeDataSource
 from laktory.models.datasources import TableDataSource
 from laktory.models.pipeline.pipelinechild import PipelineChild
-from laktory.typing import AnyDataFrame
+from laktory.typing import AnyFrame
 
 logger = get_logger(__name__)
 
@@ -99,7 +99,7 @@ class PipelineNode(BaseModel, PipelineChild):
     ```
     """
 
-    dlt_template: str = Field(
+    dlt_template: str | None = Field(
         "DEFAULT",
         description="Specify which template (notebook) to use when Delta Live Tables is selected as the orchestrator.",
     )
@@ -274,14 +274,14 @@ class PipelineNode(BaseModel, PipelineChild):
         return is_view
 
     @property
-    def stage_df(self) -> AnyDataFrame:
+    def stage_df(self) -> AnyFrame:
         """
         Dataframe resulting from reading source and applying transformer, before data quality checks are applied.
         """
         return self._stage_df
 
     @property
-    def output_df(self) -> AnyDataFrame:
+    def output_df(self) -> AnyFrame:
         """
         Dataframe resulting from reading source, applying transformer and dropping rows not meeting data quality
         expectations.
@@ -289,7 +289,7 @@ class PipelineNode(BaseModel, PipelineChild):
         return self._output_df
 
     @property
-    def quarantine_df(self) -> AnyDataFrame:
+    def quarantine_df(self) -> AnyFrame:
         """
         DataFrame storing `stage_df` rows not meeting data quality expectations.
         """
@@ -498,8 +498,8 @@ class PipelineNode(BaseModel, PipelineChild):
         apply_transformer: bool = True,
         write_sinks: bool = True,
         full_refresh: bool = False,
-        named_dfs: dict[str, AnyDataFrame] = None,
-    ) -> AnyDataFrame:
+        named_dfs: dict[str, AnyFrame] = None,
+    ) -> AnyFrame:
         """
         Execute pipeline node by: