+++ laktory/sqlparser.py
@@ -284,6 +284,7 @@ class SQLParser:
 
             decimals = expr.args.get("decimals", None)
             if decimals is not None:
+                # TODO: Fix for round(x, 3)
                 args += [decimals]
 
             args = [self.visit_expr(arg) for arg in args]